SAGE-1406: cert: add support for creating a new registration key-pair
If the `-k` option is _not_ provided a new registration key-pair
is created in addition to the CA signed certificate.

+++ ROOTFS/usr/bin/create-key-cert.sh
@@ -1,19 +1,21 @@
 #!/bin/bash -e
 
 DEFAULT_CA_PATH=./beekeeper-keys/certca/beekeeper_ca_key
-DEFAULT_KEY_PATH=./beekeeper-keys/registration_keys/registration.pub
 DEFAULT_OUT_PATH=./cert
+DEFAULT_KEY_GEN_TYPE=ed25519
+DEFAULT_REG_KEY=sage_registration
 
 print_help() {
   echo """
-usage: ${0} -b <beehive name> [-e <expire date>] [-c <CA path>] [-k <registration public key path>] [-o <outdir>]
+usage: ${0} -b <beehive name> [-e <expire date>] [-c <CA path>] [-k <registration private key path>] [-t <key type>] [-o <outdir>]
 
 Creates a node registration certificate (signed by a certificate authority).
 
   -b : the beehive the node is to be assigned upon registration (required)
   -e : (optional) certificate expire date (ex. '+1d'; see ssh-keygen(1): 'validity_interval'), certificate will be valid forever if not provided (default: valid forever)
   -c : (optional) path to the certificate authority key used in creation of the registration certificate (default: ${DEFAULT_CA_PATH})
-  -k : (optional) path to the registration public key for which the certificate is to be created (default: ${DEFAULT_KEY_PATH})
+  -k : (optional) path to an existing registration private key for which the certificate is to be created (default: create new key-pair)
+  -t : (optional) registration key type (default: ${DEFAULT_KEY_GEN_TYPE})
   -o : (optional) directory created to store output registration certificate (default: ${DEFAULT_OUT_PATH})
   -? : print this help menu
 """
@@ -22,9 +24,10 @@ Creates a node registration certificate (signed by a certificate authority).
 BEEHIVE=
 VALID="always:forever"
 CA_PATH=${DEFAULT_CA_PATH}
-KEY_PATH=${DEFAULT_KEY_PATH}
+KEY_PATH=
+KEY_GEN_TYPE=${DEFAULT_KEY_GEN_TYPE}
 OUT_PATH=${DEFAULT_OUT_PATH}
-while getopts "b:e:c:k:o:n?" opt; do
+while getopts "b:e:c:k:t:o:n?" opt; do
     case $opt in
         b) BEEHIVE=${OPTARG}
             ;;
@@ -33,6 +36,8 @@ while getopts "b:e:c:k:o:n?" opt; do
         c) CA_PATH=${OPTARG}
             ;;
         k) KEY_PATH=${OPTARG}
+            ;;
+        t) KEY_GEN_TYPE=${OPTARG}
             ;;
         o) OUT_PATH=${OPTARG}
             ;;
@@ -49,9 +54,9 @@ if [ ! -f "${CA_PATH}" ]; then
     exit 1
 fi
 
-# validate the Registration key is found
-if [ ! -f "${KEY_PATH}" ]; then
-    echo "Error: registration public key [${KEY_PATH}] not found. Exiting."
+# validate the key gen type is not empty
+if [ -z "${KEY_GEN_TYPE}" ]; then
+    echo "Error: key type must not be empty. Exiting"
     exit 1
 fi
 
@@ -61,10 +66,25 @@ if [[ ! $BEEHIVE =~ ^[A-Za-z0-9_\-]+$ ]]; then
     exit 1
 fi
 
-# make output directory and copy registration public and private key
+## create the registration certificate
 mkdir -p ${OUT_PATH}
-cp ${KEY_PATH} ${OUT_PATH}/
-cp ${KEY_PATH%.*} ${OUT_PATH}/
+
+# create registration key if one is not provided
+if [ -z "${KEY_PATH}" ]; then
+    echo "Generating new registration key-pair [type: ${KEY_GEN_TYPE}]."
+    REGPATH=${OUT_PATH}/${DEFAULT_REG_KEY}
+    ssh-keygen -f ${REGPATH} -t ${KEY_GEN_TYPE} -N ''
+else
+    # validate the Registration key is found
+    if [ -f "${KEY_PATH}" ]; then
+        cp ${KEY_PATH} ${OUT_PATH}/
+        cp ${KEY_PATH}.pub ${OUT_PATH}/
+        REGPATH=${OUT_PATH}/$(basename ${KEY_PATH})
+    else
+        echo "Error: registration private key [${KEY_PATH}] not found. Exiting."
+        exit 1
+    fi
+fi
 
 # create the certificate
 ssh-keygen \
@@ -78,6 +98,6 @@ ssh-keygen \
     -O no-user-rc \
     -O no-x11-forwarding \
     -O force-command="/opt/sage/beekeeper/register/register.sh -b ${BEEHIVE}" \
-    ${OUT_PATH}/$(basename ${KEY_PATH})
+    ${REGPATH}
 
 echo "Registration certificate created successfully! [${OUT_PATH}]"
